requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
numpy>=1.26.0
faster-whisper>=1.1.0
webrtcvad-wheels>=2.0.14
prometheus-client>=0.21.0
websockets>=12.0
python-dotenv>=1.0.0
soundfile>=0.12.0
